worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery==5.3.4
redis==5.0.1
pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
trimesh==4.0.10
numpy==1.26.3
httpx==0.27.0
python-dotenv==1.0.0
supabase==2.7.4
websockets==15.0.1
rembg==2.0.56
transformers==4.57.1
accelerate==1.11.0
diffusers==0.35.2
einops==0.8.1
omegaconf==2.3.0
safetensors==0.6.2
huggingface-hub==0.36.0
xatlas==0.0.9
imageio[ffmpeg]==2.37.2
PyMCubes==0.1.6
